Move template serving endpoint code to a common class
This reduces code duplication and makes it easier to add additional
templated endpoints in the future.
